requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
click==8.2.1
distro==1.9.0
dnspython==2.7.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.95.1
pdf2image==1.17.0
pillow==11.3.0
pydantic==2.11.7
pydantic-core==2.33.2
pymongo==4.13.2
python-dotenv==1.1.1
python-multipart==0.0.20
redis==6.2.0
rq==2.4.0
resend==0.7.2
sniffio==1.3.1
starlette==0.47.1
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
uvicorn==0.35.0
